% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicators}
\alias{indicators}
\title{Sample data with 10 economic indicators}
\format{
A data frame with 263 rows and 11 variables:

\describe{
  \item{date}{Dates}
  \item{total_production}{Industrial production in Israel - total}
  \item{retail_revenue}{ }
  \item{services_revenue}{}
  \item{employment}{}
  \item{export_services}{}
  \item{building_starts}{}
  \item{import_consumer_goods}{}
  \item{import_production_inputs}{}
  \item{export_goods}{ }
  \item{job_openings}{}
}
}
\source{
Public data
}
\usage{
indicators
}
\description{
A dataset containing 10 monthly economic indicators. All variables are log-differenced, if neccesary, to achieve stationarity.
}
\keyword{datasets}
