% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wex.R
\name{wex}
\alias{wex}
\title{Exact observation weights for the Kalman filter and smoother}
\usage{
wex(a0 = NULL, P0 = NULL, Tt, Zt, HHt, GGt, yt, t, package = "FKF")
}
\arguments{
\item{a0}{A numeric vector specifying the initial state estimate. Defaults to a vector of zeros.}

\item{P0}{A numeric matrix specifying the covariance matrix of the initial state. Defaults to a diagonal matrix with large values (e.g., 1e6) on the diagonal.}

\item{Tt}{An array specifying the transition matrix of the state equation (see \bold{Details}).}

\item{Zt}{An array specifying the observation matrix of the measurement equation (see \bold{Details}).}

\item{HHt}{An array specifying the covariance matrix of the state disturbances (see \bold{Details}).}

\item{GGt}{An array specifying the covariance matrix of the observation disturbances (see \bold{Details}).}

\item{yt}{An \eqn{d \times n}{d * n} matrix of observations, where d is the dimension and n is the number of observations.  Missing values (\code{NA}) are allowed.}

\item{t}{An integer specifying the time index at which the observation weights are evaluated.}

\item{package}{A character string indicating which backend to use (\code{"FKF"} or \code{"KFAS"}). Defaults to \code{"FKF"}.}
}
\value{
A list with two components:
\itemize{
\item \code{Wt}: An array of filtering weights with dimensions \eqn{m \times d \times n}, , where \eqn{m} is the state
dimension, \eqn{d} is the observation dimension, and \eqn{n} is the number of time points.
\item \code{WtT}: An array of smoothing weights with the same dimensions as \code{Wt}.
}
}
\description{
Computes the exact observation weights for the Kalman filter and smoother,
following Koopman and Harvey (2003). The implementation in \code{wex}
builds on the functionality provided by the \code{FKF} and \code{KFAS}
packages. These packages rely on different computational approaches:
\code{FKF} uses routines from BLAS and LAPACK, whereas \code{KFAS} uses
sequential processing, which allows the prediction error variance matrices
to be singular.
}
\details{
\strong{State space form}

\deqn{\alpha_{t+1} = T_t \alpha_t + H_t \eta_t}
\deqn{y_t = Z_t \alpha_t + G_t \epsilon_t}

where \eqn{y_t} represents the observed data (possibly with NA's),
and \eqn{\alpha_t} is the state vector.
}
\examples{

# Decompose a local level model (Nile data set)
data(Nile)
y <- Nile
wts <- wex(Tt=matrix(1),
Zt=matrix(1),
HHt = matrix(1385.066),
GGt = matrix(15124.13),
yt = t(y),
t=50)





}
\references{
Koopman, S. J., and Harvey, A. (2003). Computing observation weights for
signal extraction and filtering. \emph{Journal of Economic Dynamics and Control},
\bold{27}(7), 1317-1333.

Helske, J. (2017). KFAS: Exponential family state space models in R.
\emph{Journal of Statistical Software}, \bold{78}, 1-39.
}
\author{
Tim Ginker
}
