% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wex.R
\name{wex}
\alias{wex}
\title{Exact observation weights for the Kalman filter and smoother.}
\usage{
wex(Tt, Zt, HHt, GGt, yt, t)
}
\arguments{
\item{Tt}{An \code{array} giving the factor of the transition equation (see \bold{Details}).}

\item{Zt}{An \code{array} giving the factor of the measurement equation (see \bold{Details}).}

\item{HHt}{An \code{array} giving the variance of the innovations of the transition equation (see \bold{Details}).}

\item{GGt}{An \code{array} giving the variance of the disturbances of the measurement equation (see \bold{Details}).}

\item{yt}{An \eqn{n \times d}{n * d} matrix, where d is the dimension and n is the number of observations. \code{matrix} containing the observations. “NA”-values are allowed (see \bold{Details}).}

\item{t}{An observation index for which the weights are returned.}
}
\value{
weight matrices for filtering and smoothing.
}
\description{
The function computes exact observation weights for the Kalman filter and smoother,
as described in Koopman and Harvey (2003). 'wex' was built upon the existing 'FKF' package.
}
\details{
\strong{State space form}
}
\references{
Koopman, S. J., & Harvey, A. (2003). Computing observation weights for
 signal extraction and filtering. \emph{Journal of Economic Dynamics and Control}, \bold{27}(7), 1317-1333.
}
